refactor: features/auth 리팩토링
