+++ src/features/auth/ui/LoginKakaoButton.tsx
@@ -1,0 +1,11 @@
+import Image from 'next/image'
+import { useOauthLogin } from '../lib/useOauthLogin'
+
+export const LoginKakaoButton = () => {
+  const { handleKakaoLogin } = useOauthLogin()
+  return (
+    <div className="mt-4 mb-10 cursor-pointer" onClick={handleKakaoLogin}>
+      <Image width={340} height={52} src={'/images/icon/kakaoBtn.svg'} alt="카카오톡-로그인-버튼" />
+    </div>
+  )
+}
